% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{palette_ui}
\alias{palette_ui}
\alias{palette_c}
\alias{palette_d}
\title{Palette UI}
\usage{
palette_c(n, type = "warm", ...)

palette_d(n, type = "div", ...)
}
\arguments{
\item{n}{Describe n.}

\item{type}{Describe type.}

\item{...}{Describe dots.}
}
\description{
Palette UI
}
