% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_ui}
\alias{scale_ui}
\alias{fill_c}
\alias{colour_c}
\alias{color_c}
\alias{fill_b}
\alias{colour_b}
\alias{color_b}
\alias{fill_d}
\alias{colour_d}
\alias{color_d}
\title{Scale UI}
\usage{
fill_c(type = "warm", ...)

colour_c(type = "warm", ...)

color_c(type = "warm", ...)

fill_b(type = "warm", ...)

colour_b(type = "warm", ...)

color_b(type = "warm", ...)

fill_d(type = "qual", ...)

colour_d(type = "div", ...)

color_d(type = "div", ...)
}
\arguments{
\item{type}{Describe type.}

\item{...}{Describe dots.}
}
\description{
A more traditional UI is available through internal functions if this
would make more sense to you.
}
